Pre-check for lback directory on development install

+++ setup.py
@@ -52,7 +52,8 @@ class installsetup(install):
       create_db()
       base_dir = "%s/.lback"%(os.getenv("HOME"))
       backup_dir = "%s/backups"%( base_dir )
-      shutil.rmtree( base_dir )
+      if os.path.exists( base_dir ):
+         shutil.rmtree( base_dir )
       os.makedirs( base_dir )
       os.makedirs( backup_dir )
       shutil.copy( "./settings.json", "%s/%s"%( base_dir, "settings.json" ) )
